remove stub _other.go files
