--- pkg/utils/erofs/erofs_other.go
@@ -1,20 +1,0 @@
-//go:build !linux
-// +build !linux
-
-package erofs
-
-import "fmt"
-
-const GlobalDomainName = ""
-
-func Mount(bootstrapPath, fsID, mountPoint string) error {
-	return fmt.Errorf("not implemented")
-}
-
-func Umount(mountPoint string) error {
-	return fmt.Errorf("not implemented")
-}
-
-func FscacheID(snapshotID string) string {
-	panic("not implemented")
-}
